Fix ticket #357 LinkToView cannot be displayed on Statusbar
Analysis：
In revision 1733, this.domObj.appendChild(this.anchor) was removed from the initializeWidget function and added to the setUiObj function. if the LinkToView? widget is added to StatusBar, it will not be assigned any uiClass, so the setUiObj function will never be called. That's why the widget can not be displayed on StatusBar.

Resolution:
If the widget has domObj at initialize time, we just append it to the domObj, else wait until setUiObj is called.