Fixed ticket #359 Theme on raster layer cause an error.
[Analysis]
For layers which don't contain the geometry property, the layer's name will not be record in layerNames in theme.php
So in thememain.php the layerSelect.value will be null, and the layerInfo request will fail.


[Resolution/Workaround]
To resolve this problem, I checked the layerSelect.value, if it's null, just return.